+++ wizard/spec-generator/src/components/assets/AssetCreateDialog.tsx
@@ -15,7 +15,6 @@ import { Input } from "../ui/input";
 import { Field, FieldLabel } from "../ui/field";
 import { ArrowLeft, File, Folder, PackageIcon, UploadIcon } from "lucide-react";
 import { AssetTile } from "./AssetTile";
-import { UploadedFilePreview } from "./UploadedFilePreview";
 import {
   ALLOWED_EXTENSIONS,
   getAssetName,
@@ -28,12 +27,14 @@ import {
   type AssetEntry,
   type AssetSelection,
 } from "./assetPaths";
+import { FileTypeIcon } from "./FileTypeIcon";
+import { USE_NN_BELOW } from "./assetPaths";
 
 /**
  * Dialog to pick the content of an asset: either a file uploaded by the user (custom asset)
  * or a file from the bundled assets shipped with the application.
  */
-export function AssetSelectorDialog({
+export function AssetCreateDialog({
   open,
   setOpen,
   onSelect,
@@ -250,5 +251,41 @@ export function AssetSelectorDialog({
         </DialogFooter>
       </DialogContent>
     </Dialog>
+  );
+}
+
+function UploadedFilePreview({ file }: { file: File }) {
+  const [objectUrl, setObjectUrl] = React.useState<string | null>(null);
+  const [pixelated, setPixelated] = React.useState(false);
+
+  React.useEffect(() => {
+    if (!file.type.startsWith("image/")) {
+      setObjectUrl(null);
+      return;
+    }
+    const url = URL.createObjectURL(file);
+    setObjectUrl(url);
+    return () => URL.revokeObjectURL(url);
+  }, [file]);
+
+  return (
+    <div className="flex min-h-0 w-full flex-1 flex-col items-center justify-center gap-2">
+      {objectUrl ? (
+        <img
+          src={objectUrl}
+          alt=""
+          onLoad={(e) => {
+            const img = e.currentTarget;
+            setPixelated(img.naturalWidth < USE_NN_BELOW || img.naturalHeight < USE_NN_BELOW);
+          }}
+          className={`min-h-0 w-auto max-w-full flex-1 object-contain ${
+            pixelated ? "[image-rendering:pixelated]" : ""
+          }`}
+        />
+      ) : (
+        <FileTypeIcon filePath={file.name} />
+      )}
+      <span className="max-w-full shrink-0 truncate text-sm">{file.name}</span>
+    </div>
   );
 }
